Moved arcjet instantiation and event transform to a server-only-module Hooks file filters out routes the will override default Shield protection Updated "how it works" section
