+++ examples/sveltekit/README.md
@@ -44,10 +44,10 @@ This example shows how to use Arcjet to protect [SvelteKit](https://kit.svelte.d
 
 This example uses the generic [`arcjet`](https://www.npmjs.com/package/arcjet) package.
 
-The `arcjet` instance is created in `/src/hooks.server.ts`, which is configured to use [Shield](https://docs.arcjet.com/shield) for all protected requests. The hooks file also sets up an object containing the request information required by arcjet's `protect()` method.
+The `arcjet` instance is created in `/src/lib/server/arcjet.ts` and is configured to enable [Shield](https://docs.arcjet.com/shield) by default. This server-only module also provides a `transformEvent()` method that takes a Svelte event and transforms it into the format required by the `arcjet` package.
 
-The home page's server-side scripts in `/src/routes/+page.server.ts` retrieve the `arcjet` instance, and call its `protect()` method, thereby protecting this page through [Shield](https://docs.arcjet.com/shield).
+`/src/hooks.server.ts` imports the `arcjet` instance and runs the `protect()` method on all requests. The only exception is any requests who's pathanme is listed in `filteredRoutes`, in which case protection is left to that route's server code.
 
-The rate-limited page also has a server-side script file at `/src/routes/rate-limited/+page.server.ts`, which retrieves the `arcjet` instance, and then applies extra rate-limiting rules before calling the `protect()` method.
+The rate-limited page has a server-side script file at `/src/routes/rate-limited/+page.server.ts`, which retrieves the `arcjet` instance, and then applies extra rate-limiting rules before calling the `protect()` method.
 
 Finally, the rate-limited API end-point performs the same augmentation of the rules as the rate-limted web page, as can be seen in `/src/routes/api/rate-limited/+server.ts`.
